Made modifications to deliberation cycle (and related parts) to spend less time processing agents outside the deliberation cycle.
Instead of copying the set of scheduled agents in each time step, we now replace
that queue by reference, which should be O(1)

Instead of putting agents in a hasmap, we let the deliberation cycle of each agent
return an object that contains both the agentID and the actions, and return a list
of Futures of those objects. THis has two advantages:
1. We do not have to iterate over all agents and their actions anymore, until we
start actually processing these results in the environment. That loop is necessary
anyway, so we get rid of one loop
2. Populating a hashmap can be expensive, if buckets need to be recalculated.

The disadvantage is a n more object creations for n agents, but this happens within
the deliberation cycle, so can benefit from the multiple threads available, resulting
in better scaling (hopefully)